requirements.txt
mediapipe
opencv-python
face_recognition
ultralytics
torch
numpy
pillow